+++ src/components/department/department-management.tsx
@@ -8,8 +8,9 @@ import { Badge } from "@/components/ui/badge"
 import { Avatar, AvatarFallback } from "@/components/ui/avatar"
 import { Building2, Edit, Eye, MoreHorizontal, Plus, Search, Trash2, Users } from "lucide-react"
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from "@/components/ui/dropdown-menu"
-import AddDepartmentModal from "./add-department-modal"
-import DepartmentDetailModal from "./department-detail-modal"
+import dynamic from "next/dynamic"
+const AddDepartmentModal = dynamic(() => import("./add-department-modal"), { ssr: false })
+const DepartmentDetailModal = dynamic(() => import("./department-detail-modal"), { ssr: false })
 import { useState } from "react"
 import LoadingSpinner from "@/components/ui/loading-spinner"
 import { useSWRConfig } from "swr"
@@ -17,7 +18,7 @@ import { deleteDepartment } from "@/services/department"
 import ConfirmDeleteModal from "../shared/confirm-delete-modal"
 import { toast } from "sonner"
 import { getPositionValue } from "@/lib/utils"
-import DepartmentTable from "./DepartmentTable"
+const DepartmentTable = dynamic(() => import("./DepartmentTable"), { ssr: false })
 
 export default function DepartmentManagement({ departments, summary, isLoading }: { departments: any[], summary: any, isLoading: boolean }) {
   const { mutate } = useSWRConfig()
