+++ src/types/auth.ts
@@ -1,0 +1,8 @@
+export interface LoginResponse {
+  token: string;
+  user: {
+    id: number;
+    email: string;
+    // Add other user fields as needed
+  };
+}
