+++ src/types/loading.ts
@@ -1,0 +1,5 @@
+export interface LoadingContextType {
+  isLoading: boolean;
+  showLoading: () => void;
+  hideLoading: () => void;
+}
